exp_disponibilidad/logs-service/requirements.txt

Flask-RESTful==0.3.9
flask
flask_cors
aiven-client
python-dotenv
eventlet
pika
flask_coney
